+++ azplugins/module.cc
@@ -14,6 +14,12 @@ namespace py = pybind11;
 #include "PairPotentials.h"
 #include "WallPotentials.h"
 
+/* Updaters */
+#include "TypeUpdater.h"
+#ifdef ENABLE_CUDA
+#include "TypeUpdaterGPU.h"
+#endif // ENABLE_CUDA
+
 //! Plugins for soft matter
 namespace azplugins
 {
@@ -75,6 +81,12 @@ PYBIND11_PLUGIN(_azplugins)
     azplugins::detail::export_pair_potential<azplugins::detail::PairEvaluatorAshbaugh>(m, "PairPotentialAshbaugh");
     azplugins::detail::export_pair_potential<azplugins::detail::PairEvaluatorColloid>(m, "PairPotentialColloid");
     azplugins::detail::export_ashbaugh_params(m);
+
+    /* Updaters */
+    azplugins::detail::export_TypeUpdater(m);
+    #ifdef ENABLE_CUDA
+    azplugins::detail::export_TypeUpdaterGPU(m);
+    #endif // ENABLE_CUDA
 
     /* Wall potentials */
     azplugins::detail::export_wall_potential<azplugins::detail::WallEvaluatorColloid>(m, "WallPotentialColloid");